requirements.txt
streamlit
requests
pydantic
wordcloud
matplotlib
python-dotenv
